requirements.txt
git+https://github.com/openai/CLIP.git
colorama==0.4.4
colorlog==4.7.2
hyperopt==0.2.7
lmdb==1.2.1
numpy==1.21.2
pandas==1.4.1
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.1
tensorboardX==2.6
torch==1.10.2
torch_geometric==2.0.4
torchvision==0.11.3
tqdm==4.63.0
transformers==4.16.2
wandb==0.13.10
